Button to all configuration types for messaging
