+++ clients/admin-ui/src/features/privacy-requests/configuration/TwilioEmailConfiguration.tsx
@@ -11,6 +11,8 @@ import {
   useGetMessagingConfigurationDetailsQuery,
 } from "~/features/privacy-requests/privacy-requests.slice";
 
+import TestMessagingProviderConnectionButton from "./TestMessagingProviderConnectionButton";
+
 const TwilioEmailConfiguration = () => {
   const [configurationStep, setConfigurationStep] = useState("");
   const { successAlert } = useAlert();
@@ -52,7 +54,7 @@ const TwilioEmailConfiguration = () => {
       handleError(result.error);
     } else {
       successAlert(`Twilio email secrets successfully updated.`);
-      setConfigurationStep("configureTwilioEmailSecrets");
+      setConfigurationStep("testConnection");
     }
   };
 
@@ -148,6 +150,11 @@ const TwilioEmailConfiguration = () => {
             </Formik>
           </Stack>
         </>
+      ) : null}
+      {configurationStep === "testConnection" ? (
+        <TestMessagingProviderConnectionButton
+          messagingDetails={messagingDetails}
+        />
       ) : null}
     </Box>
   );
